change focus to install driver and cuda only

+++ uninstall.sh
@@ -1,12 +1,16 @@
 #!/bin/bash
 
-echo "[CUDA-TSFLOW] Removing driver.."
+echo "[AUTO-CUDA] Removing driver.."
 sudo apt purge nvidia*
 sudo apt autoremove
 sudo apt autoclean
 
-echo "[CUDA-TSFLOW] Removing cuda.."
+echo "[AUTO-CUDA] Removing cuda.."
 sudo rm -rf /usr/local/cuda*
 
-echo "[CUDA-TSFLOW] Restoring linux enviroment before cuda enviroment added.."
-cp ~/.bashrc.backup.cuda ~/.bashrc
+echo "[AUTO-CUDA] Restoring linux enviroment before cuda enviroment added.."
+if [[ -f ~/.bashrc.backup.cuda ]]; then
+	cp ~/.bashrc.backup.cuda ~/.bashrc
+else
+	echo "[AUTO-CUDA] There's no bashrc backup found.."
+if
